+++ src/main/java/insane96mcp/insanesurvivaloverhaul/module/ISOModules.java
@@ -10,6 +10,7 @@ import net.neoforged.neoforge.common.ModConfigSpec;
 public class ISOModules {
 	public static final String COMBAT = InsaneSO.MOD_ID + ":combat";
 	public static final String ITEMS = InsaneSO.MOD_ID + ":items";
+	public static final String MINING = InsaneSO.MOD_ID + ":mining";
 	public static final String MISC = InsaneSO.MOD_ID + ":misc";
 	public static final String MOBS = InsaneSO.MOD_ID + ":mobs";
 	public static final String WORLD = InsaneSO.MOD_ID + ":world";
@@ -17,6 +18,7 @@ public class ISOModules {
 	public static void init(IEventBus eventBus, ModConfigSpec.Builder builder) {
 		create(COMBAT, "Combat", eventBus, builder);
 		create(ITEMS, "Items", eventBus, builder);
+		create(MINING, "Mining", eventBus, builder);
 		create(MISC, "Misc", eventBus, builder);
 		create(MOBS, "Mobs", eventBus, builder);
 		create(WORLD, "World", eventBus, builder);